sys - Move all Str function helper under a util folder
